// src/model/publish/column.model.ts
// Generated by https://quicktype.io

export interface IColumn {
	accept_submission:  boolean;
	title:              string;
	url:                string;
	comment_permission: string;
	author:             Author;
	updated:            number;
	intro:              string;
	image_url:          string;
	followers:          number;
	type:               string;
	id:                 string;
	articles_count:     number;
}

export interface Author {
	avatar_url_template: string;
	name:                string;
	headline:            string;
	gender:              number;
	user_type:           string;
	url_token:           string;
	is_advertiser:       boolean;
	avatar_url:          string;
	url:                 string;
	type:                string;
	badge:               any[];
	id:                  string;
	is_org:              boolean;
}